requirements.txt
psycopg2==2.7.5
pytest-pgsql==1.1.1
pytest-mock==1.10.0
sqlalchemy==1.2.10
xlrd==1.1.0
